Added importer support for the new Runway model. Also added two new runtime features - the ability to pass in a dict of which records should be imported, and a max # of records to import. Max records is useful for debugging or when you just need a bit of data, the dict config will be useful if you don't need a particular model's data.

+++ faadata/airports/load.py
@@ -15,30 +15,74 @@ def clean_chars(value):
     value = value.replace('\xfc', ' ')
     return value
 
-def airport_import(importfile):
-    Remark.objects.all().delete() # there's no way to version them. Start with a clean slate each time.
-    Attendance.objects.all().delete()
+def airport_import(importfile, config={'import_att': True, 'import_rmk': True, 'import_rwy': True, 'import_apt': True}, max_records=10000000):
+    if config['import_att']:
+        Attendance.objects.all().delete()
+    if config['import_rmk']:
+        Remark.objects.all().delete() # there's no way to version them. Start with a clean slate each time.
 
+    count = 0
     for line in importfile:
         data = parse_apt_line(clean_chars(line))
-        if data['record_type'] == 'ATT':
+
+        try:
+            airport = Airport.objects.get(facility_site_number=data['facility_site_number'])
+        except Airport.DoesNotExist:
+            airport = Airport(facility_site_number=data['facility_site_number'])
+
+        if data['record_type'] == 'ATT' and config['import_att']:
             try:
                 attendance = Attendance.objects.create(airport=airport, sequence=data['attendace_schedule_sequence'], schedule=data['attendance_schedule'])
             except:
                 print data
 
-        elif data['record_type'] == 'RMK':
+        elif data['record_type'] == 'RMK' and config['import_rmk']:
             try:
                 remark = Remark.objects.create(airport=airport, element_name=data['element_name'], body=data['element_text'])
             except:
                 print data
 
-        elif data['record_type'] == 'APT':
+        elif data['record_type'] == 'RWY' and config['import_rwy']:
             try:
-                airport = Airport.objects.get(facility_site_number=data['facility_site_number'])
-            except Airport.DoesNotExist:
-                airport = Airport(facility_site_number=data['facility_site_number'])
+                runway = Runway.objects.get(airport=airport, runway_identification=data['runway_identification'])
+            except Runway.DoesNotExist:
+                runway = Runway(airport=airport, runway_identification=data['runway_identification'])
 
+            runway.runway_length = data['runway_length']
+            runway.runway_width = data['runway_width']
+            runway.surface_type_condition = data['surface_type_condition']
+            runway.surface_treatment = data['surface_treatment']
+            runway.pavement_classification_number = data['pavement_classification_number']
+            runway.lights_edge_intensity = data['lights_edge_intensity']
+            if getattr(data, 'base_end_lon', False) and getattr(data, 'base_end_lat', False):
+                runway.base_end_point = Point((data['base_end_lon'], data['base_end_lat']))
+            if data['base_end_elevation_physical_runway_end']:
+                runway.base_end_elevation_physical_runway_end = Decimal(data['base_end_elevation_physical_runway_end'])
+            if getattr(data, 'base_end_displaced_threshold_lon', False) and getattr(data, 'base_end_displaced_threshold_lat', False):
+                runway.base_end_displaced_threshold_point = Point((data['base_end_displaced_threshold_lon'], data['base_end_displaced_threshold_lat']))
+            if data['base_end_elevation_displaced_threshold']:
+                runway.base_end_elevation_displaced_threshold = Decimal(data['base_end_elevation_displaced_threshold'])
+            if data['base_end_displaced_threshold_length_from_end']:
+                runway.base_end_displaced_threshold_length_from_end = data['base_end_displaced_threshold_length_from_end']
+            runway.base_end_visual_glide_slope_indicators = data['base_end_visual_glide_slope_indicators']
+            runway.base_end_runway_visual_range_equipment_locations = data['base_end_runway_visual_range_equipment_locations']
+            runway.base_end_runway_visual_range_equipment = convert_boolean(data['base_end_runway_visual_range_equipment'])
+            if getattr(data, 'reciprocal_end_lon', False) and getattr(data, 'reciprocal_end_lat', False):
+                runway.reciprocal_end_point = Point((data['reciprocal_end_lon'], data['reciprocal_end_lat']))
+            if data['reciprocal_end_elevation_physical_runway_end']:
+                runway.reciprocal_end_elevation_physical_runway_end = Decimal(data['reciprocal_end_elevation_physical_runway_end'])
+            if getattr(data, 'reciprocal_end_displaced_threshold_lon', False) and getattr(data, 'reciprocal_end_displaced_threshold_lat', False):
+                runway.reciprocal_end_displaced_threshold_point = Point((data['reciprocal_end_displaced_threshold_lon'], data['reciprocal_end_displaced_threshold_lat']))
+            if data['reciprocal_end_elevation_displaced_threshold']:
+                runway.reciprocal_end_elevation_displaced_threshold = Decimal(data['reciprocal_end_elevation_displaced_threshold'])
+            if data['reciprocal_end_displaced_threshold_length_from_end']:
+                runway.reciprocal_end_displaced_threshold_length_from_end = data['reciprocal_end_displaced_threshold_length_from_end']
+            runway.reciprocal_end_visual_glide_slope_indicators = data['reciprocal_end_visual_glide_slope_indicators']
+            runway.reciprocal_end_runway_visual_range_equipment_locations = data['reciprocal_end_runway_visual_range_equipment_locations']
+            runway.reciprocal_end_runway_visual_range_equipment = convert_boolean(data['reciprocal_end_runway_visual_range_equipment'])
+            runway.save()
+
+        elif data['record_type'] == 'APT' and config['import_apt']:
             airport.location_identifier = data['location_identifier']
             airport.facility_name = data['facility_name']
             airport.ownership_type = data['ownership_type']
@@ -70,7 +114,10 @@ def airport_import(importfile):
             except:
                 print('Airport data fail for %s' % data['location_identifer'])
 
+        count += 1
+        if count > max_records:
+            exit()
 
 if __name__ == '__main__':
-    path = '/Users/adam/Downloads/56DySubscription_November_18__2010_-_January_13__2011/'
-    airport_import(open(path + 'APT.txt'))
+    path = '/Users/adam/Downloads/56DySubscription_January_13__2011_-_March_10__2011/'
+    airport_import(open(path + 'APT.txt'), {'import_att': False, 'import_rmk': False, 'import_rwy': True, 'import_apt': False}) # , max_records=100)
